# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusterhubGo",
        "description": "This api is for working with musterhabGo",
        "contact": {
            "email": "artem.yablochnyi@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Business service"
                ],
                "summary": "Services list",
                "description": "Returns services list",
                "operationId": "services",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items to return",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServiceCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/services/{service}/questions": {
            "get": {
                "tags": [
                    "Business service"
                ],
                "summary": "Questions list for given service",
                "description": "Returns questions list for given service",
                "operationId": "questions",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Business service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items to return",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServicesQuestionCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Business service"
                ],
                "summary": "Creates new question for given business service",
                "description": "Returns created question data",
                "operationId": "questions-create",
                "parameters": [
                    {
                        "name": "service",
                        "in": "path",
                        "description": "Business service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusinessServicesQuestionCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServicesQuestionResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/questions/{id}": {
            "get": {
                "tags": [
                    "Business service question"
                ],
                "summary": "Get question information",
                "description": "Returns question data by id",
                "operationId": "question-id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServicesQuestionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Business service question"
                ],
                "summary": "Updates question by id",
                "description": "Returns updated question",
                "operationId": "question-update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BusinessServicesQuestionUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessServicesQuestionResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Business service question"
                ],
                "summary": "Deletes question by id",
                "description": "Deletes question by id",
                "operationId": "question-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BusinessService": {
                "title": "BusinessService",
                "description": "BusinessService model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "business_id": {
                        "title": "business_id",
                        "description": "Related business profile id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Service title",
                        "type": "string"
                    },
                    "price": {
                        "$ref": "#/components/schemas/Price"
                    },
                    "description": {
                        "title": "description",
                        "description": "Service description",
                        "type": "string",
                        "nullable": true
                    },
                    "duration": {
                        "title": "duration",
                        "description": "Duration in minutes",
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "is_field": {
                        "title": "is_field",
                        "description": "Indicates whether the service can be provided in the fields",
                        "type": "boolean"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "Owner id",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Date and time of creation",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-05-28T13:51:36.000000Z"
                    },
                    "updated_at": {
                        "title": "description",
                        "description": "Date and time of updation",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-05-28T13:51:36.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessService"
                }
            },
...
            "BusinessServicesQuestion": {
                "title": "BusinessServicesQuestion",
                "description": "BusinessServicesQuestion model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "service_id": {
                        "title": "service_id",
                        "description": "Related business service id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "question": {
                        "title": "question",
                        "description": "Question text",
                        "type": "string"
                    },
                    "answer": {
                        "title": "answer",
                        "description": "Answer text",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "Date and time of creation",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-05-28T13:51:36.000000Z"
                    },
                    "updated_at": {
                        "title": "description",
                        "description": "Date and time of updation",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-05-28T13:51:36.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessServicesQuestion"
                }
            },
            "Currency": {
                "title": "Currency",
                "description": "Currency model. Implements country currency codes according to ISO 4217 standard",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "alpha_code": {
                        "title": "Alpha code",
                        "description": "Currency alphabetical code",
                        "type": "string",
                        "example": "RUB"
                    },
                    "numeric_code": {
                        "title": "Numeric code",
                        "description": "Currency numeric code",
                        "type": "string",
                        "example": "643"
                    },
                    "fraction": {
                        "title": "Fraction",
                        "description": "Number of fractional units (e.g. in dollar 100 cents)",
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0,
                        "multipleOf": 10,
                        "example": 100
                    },
                    "sign": {
                        "title": "Sign",
                        "description": "Currency sign",
                        "type": "string",
                        "example": "₽",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Currency"
                }
            },
...
            "BusinessServicesQuestionCreateRequest": {
                "title": "BusinessServicesQuestionCreateRequest",
                "description": "Creates question",
                "required": [
                    "question",
                    "answer"
                ],
                "properties": {
                    "question": {
                        "title": "question",
                        "description": "Text of the question",
                        "type": "string",
                        "maximum": 3000,
                        "minimum": 5
                    },
                    "answer": {
                        "title": "answer",
                        "description": "Text of the answer",
                        "type": "string",
                        "maximum": 3000,
                        "minimum": 5
                    }
                },
                "type": "object"
            },
            "BusinessServicesQuestionUpdateRequest": {
                "title": "BusinessServicesQuestionUpdateRequest",
                "description": "Updates question",
                "properties": {
                    "question": {
                        "title": "question",
                        "description": "Text of the question",
                        "type": "string",
                        "maximum": 3000,
                        "minimum": 5
                    },
                    "answer": {
                        "title": "answer",
                        "description": "Text of the answer",
                        "type": "string",
                        "maximum": 3000,
                        "minimum": 5
                    }
                },
                "type": "object"
            },
...
            "BusinessServiceCollection": {
                "title": "BusinessServiceCollection",
                "description": "BusinessService collection",
                "properties": {
                    "count": {
                        "title": "Count",
                        "description": "Number of items in response",
                        "type": "integer",
                        "example": 1
                    },
                    "items": {
                        "title": "Items",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BusinessService"
                        }
                    },
                    "total": {
                        "title": "Total",
                        "description": "Number of items in database",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessServiceCollection"
                }
            },
...
            "BusinessServicesQuestionCollection": {
                "title": "BusinessServicesQuestionCollection",
                "description": "BusinessServices question collection",
                "properties": {
                    "count": {
                        "title": "Count",
                        "description": "Number of items in response",
                        "type": "integer",
                        "example": 1
                    },
                    "items": {
                        "title": "Items",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BusinessServicesQuestion"
                        }
                    },
                    "total": {
                        "title": "Total",
                        "description": "Number of items in database",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessServicesQuestionCollection"
                }
            },
            "BusinessServicesQuestionResource": {
                "title": "BusinessServicesQuestionResource",
                "description": "BusinessService question resource",
                "properties": {
                    "item": {
                        "$ref": "#/components/schemas/BusinessServicesQuestion"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessServicesQuestionResource"
                }
            },
...
            "UnprocessableContentResponse": {
                "title": "UnprocessableContentResponse",
                "description": "Response when validation failed",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Error message",
                        "type": "string"
                    },
                    "errors": {
                        "$ref": "#/components/schemas/UnprocessableContentResponseError"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UnprocessableContentResponse"
                }
            },
            "UnprocessableContentResponseError": {
                "title": "Error",
                "description": "Error",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Faild field name",
                        "type": "array",
                        "items": {
                            "description": "Error message",
                            "type": "string"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Error"
                }
            },
            "Price": {
                "title": "Price",
                "description": "Price",
                "properties": {
                    "amount": {
                        "title": "amount",
                        "description": "Amount",
                        "type": "number",
                        "example": 129.99
                    },
                    "currency": {
                        "$ref": "#/components/schemas/Currency"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Price"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}